refactor(anthropic): group message cache-control helpers in cache_control.go
Pure move: messageCacheBreakpoints, applyMessageCacheControl, and
applyBetaMessageCacheControl relocate unchanged so the prompt-caching
breakpoint logic lives in one file.
