+++ pkg/model/provider/anthropic/client.go
@@ -444,17 +444,6 @@ func (c *Client) convertMessagesWithDeferred(ctx context.Context, messages []cha
 	return anthropicMessages, nil
 }
 
-// messageCacheBreakpoints returns how many message-level cache_control
-// breakpoints a request may use. Anthropic allows at most 4 per request;
-// deferred tools consume one on the tool list, leaving one fewer for
-// messages.
-func messageCacheBreakpoints(requestTools []tools.Tool) int {
-	if containsDeferredTool(requestTools) {
-		return 1
-	}
-	return 2
-}
-
 func deferredToolNamesByCallID(requestTools []tools.Tool) map[string][]string {
 	result := make(map[string][]string)
 	for _, tool := range requestTools {
@@ -631,21 +620,6 @@ func (c *Client) convertUserMultiContent(ctx context.Context, parts []chat.Messa
 	}
 
 	return contentBlocks, nil
-}
-
-// applyMessageCacheControl adds ephemeral cache control to the last content block
-// of the last `breakpoints` messages for prompt caching.
-func applyMessageCacheControl(messages []anthropic.MessageParam, breakpoints int) {
-	for i := len(messages) - 1; i >= 0 && i >= len(messages)-breakpoints; i-- {
-		content := messages[i].Content
-		if len(content) == 0 {
-			continue
-		}
-		// nil for block kinds without cache control (e.g. thinking blocks).
-		if cc := content[len(content)-1].GetCacheControl(); cc != nil {
-			*cc = anthropic.NewCacheControlEphemeralParam()
-		}
-	}
 }
 
 // extractSystemBlocks converts any system-role messages into Anthropic system text blocks
